Create draw.py
The above code takes an image as input and displays it and then we mark 6 points in the image and once completed the area inside the 6 points gets cropped. while saving the points in a json and making sure to check if points already exists in the file then it wont ask user to mark point again. and just crops the image directly